fix: アーカイブのルーティングが一部旧サイトと違ったのを修正
