+++ README.md
@@ -9,6 +9,8 @@
 
 ### 記事
 
+- **一覧で、本文の冒頭ではなく、Twitter用コメントを概要として使うようにしました**
+  - 通信量削減のためです
 - 記事投稿後、ビルドをトリガーする必要がなくなり、待ち時間がゼロになりました
   - 公開直後は404になる場合があります
   - 一覧の2ページ目以降の更新がすぐに行われない場合があります
