+++ src/app/articles/category/[categoryId]/[page]/page.tsx
@@ -1,26 +1,24 @@
 import {
   getAllArticleIds,
-  getAllSerieses,
+  getAllCategories,
   getArticles,
-  getSeries,
+  getCategory,
 } from '@/utils/micro-cms';
 import { Metadata } from 'next';
 import ArticleList from '@/components/ArticleList';
-import { clientEnv } from '@/utils/client-env';
+import { clientEnv } from '@/config/client-env';
 import { notFound } from 'next/navigation';
 
-export const revalidate = 3600;
-
 type Params = {
-  params: { seriesId: string; page: string };
+  params: { categoryId: string; page: string };
 };
 
 export async function generateStaticParams() {
-  const serieses = await getAllSerieses();
+  const categories = await getAllCategories();
   const paths = await Promise.all(
-    serieses.contents
-      .map(async ({ id: seriesId }) => {
-        return await getAllArticleIds({ seriesId }).then((articles) => {
+    categories.contents
+      .map(async ({ id: categoryId }) => {
+        return await getAllArticleIds({ categoryId }).then((articles) => {
           // 必要なページ数を計算
           const pages = articles
             ? Math.ceil(
@@ -30,7 +28,7 @@ export async function generateStaticParams() {
           return Array.from({ length: pages }, (_, i) =>
             (i + 1).toString()
           ).map((page) => ({
-            seriesId,
+            categoryId,
             page,
           }));
         });
@@ -43,29 +41,32 @@ export async function generateStaticParams() {
 export async function generateMetadata({
   params,
 }: Params): Promise<Metadata | void> {
-  const series = await getSeries(params.seriesId);
-  if (series) {
+  const category = await getCategory(params.categoryId);
+  if (category) {
     return {
-      title: `${series.series}の記事一覧 ${params.page}ページ目`,
+      title: `${category.category}の記事一覧 ${params.page}ページ目`,
     };
   }
 }
 
-export default async function ArticleSeriesIndexPage({
-  params: { seriesId, page },
+export default async function ArticleCategoryIndexPage({
+  params: { categoryId, page },
 }: Params) {
   const pageNumber = page ? Number(page) : 1;
   const articles = await getArticles(pageNumber, {
-    seriesId,
+    categoryId,
   });
-  if (!articles || articles.contents.length === 0) {
+  if (!articles || articles?.contents.length === 0) {
     notFound();
   }
+
   return (
-    <ArticleList
-      data={articles}
-      pageNumber={pageNumber}
-      paginationPath={`/articles/series/${seriesId}`}
-    />
+    <>
+      <ArticleList
+        data={articles}
+        pageNumber={pageNumber}
+        paginationPath={`/articles/category/${categoryId}`}
+      />
+    </>
   );
 }
