+++ src/app/members/not-found.tsx
@@ -1,0 +1,5 @@
+import ErrorInfo from '@/components/ErrorInfo';
+
+export default function NotFound() {
+  return <ErrorInfo title="ページが見つかりませんでした"></ErrorInfo>;
+}
